fuzzy number and works on double

+++ app/src/main/java/com/example/fuzzynumbercomparator/fuzzynumber/ComparingStrategy.kt
@@ -1,5 +1,5 @@
 package com.example.fuzzynumbercomparator.fuzzynumber
 
 interface ComparingStrategy {
-    fun compare(firstNumber: FuzzyNumber, secondNumber: FuzzyNumber): Float
+    fun compare(firstNumber: FuzzyNumber, secondNumber: FuzzyNumber): Double
 }
